+++ keccak256/Cargo.toml
@@ -8,3 +8,5 @@ halo2 = "0.0"
 tiny-keccak = {version="2.0.2", features = ["keccak"]}
 pasta_curves = "0.1"
 itertools = "0.10.1"
+num-bigint = "0.4.2"
+num-traits = "0.2.14"
